build_docs continued. windows

+++ build_docs/macosdocs.sh
@@ -5,7 +5,7 @@
 # Distributed under the Boost Software License, Version 1.0.
 # (See accompanying file LICENSE_1_0.txt or copy at http://boost.org/LICENSE_1_0.txt)
 
-set -e
+set -ex
 shopt -s extglob
 shopt -s dotglob
 
@@ -160,12 +160,15 @@ fi
 # Generally, boostorg/release-tools treats all libraries the same, meaning it installs one set of packages and executes b2.
 # Therefore all libraries ought to build under 'main' and shouldn't need anything customized.
 
-all_types="main cppalv1"
+all_types="main antora cppalv1"
 # cppalv1_types="json beast url http_proto socks_proto zlib"
 cppalv1_types="not_currently_used skipping_this"
 
 if [ -z "$typeoption" ]; then
-    if [[ " $cppalv1_types " =~ .*\ $REPONAME\ .* ]]; then
+    if [ -f "$BOOST_SRC_FOLDER/doc/build_antora.sh" ]; then
+        typeoption="antora"
+    elif [[ " $cppalv1_types " =~ .*\ $REPONAME\ .* ]]; then
+        # not used currently, but an example
         typeoption="cppalv1"
     else
         typeoption="main"
@@ -175,7 +178,7 @@ fi
 echo "Build type is ${typeoption}."
 
 if [[ !  " $all_types " =~ .*\ $typeoption\ .* ]]; then
-    echo "Allowed types are currently 'main' and 'cppalv1'. Not $typeoption. Please choose a different option. Exiting."
+    echo "Allowed types are currently 'main', 'antora' and 'cppalv1'. Not $typeoption. Please choose a different option. Exiting."
     exit 1
 fi
 
@@ -186,6 +189,25 @@ echo '==================================> INSTALL'
 # graphviz package added for historical reasons, might not be used.
 
 if [ "$skippackagesoption" != "yes" ]; then
+
+    if [ "$typeoption" = "antora" ]; then
+        mkdir -p ~/.nvm_${REPONAME}_antora
+        export NODE_VERSION=18.18.1
+        # The container has a pre-installed nodejs. Overwrite those again.
+        export NVM_BIN="$HOME/.nvm_${REPONAME}_antora/versions/node/v${NODE_VERSION}/bin"
+        export NVM_DIR=$HOME/.nvm_${REPONAME}_antora
+        export NVM_INC=$HOME/.nvm_${REPONAME}_antora/versions/node/v${NODE_VERSION}/include/node
+        curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.39.5/install.sh | bash
+        export NVM_DIR=$HOME/.nvm_${REPONAME}_antora
+        . "$NVM_DIR/nvm.sh" && nvm install ${NODE_VERSION}
+        . "$NVM_DIR/nvm.sh" && nvm use v${NODE_VERSION}
+        . "$NVM_DIR/nvm.sh" && nvm alias default v${NODE_VERSION}
+        export PATH="$HOME/.nvm_${REPONAME}_antora/versions/node/v${NODE_VERSION}/bin/:${PATH}"
+        node --version
+        npm --version
+        npm install gulp-cli@2.3.0
+        npm install @mermaid-js/mermaid-cli@10.5.1
+    fi
 
     brew install ruby
     sudo ln -s /opt/homebrew/opt/ruby/bin/ruby /usr/local/bin/ruby || true
@@ -253,22 +275,37 @@ if [ "$skippackagesoption" != "yes" ]; then
     # an alternate set of packages from https://www.boost.org/doc/libs/develop/doc/html/quickbook/install.html
     # sudo port install libxslt docbook-xsl docbook-xml-4.2
 
-    cd $BOOST_SRC_FOLDER
-    cd $BOOSTROOTRELPATH
-    mkdir -p tmp && cd tmp
-    if [ ! -f saxonhe.zip ]; then curl -s -S --retry 10 -L -o saxonhe.zip https://sourceforge.net/projects/saxon/files/Saxon-HE/9.9/SaxonHE9-9-1-4J.zip/download; fi
-    unzip -d saxonhe -o saxonhe.zip
-    cd saxonhe
+    if [ "$typeoption" = "cppalv1" ] || [ "$typeoption" = "main" ]; then
+        cd $BOOST_SRC_FOLDER
+        cd $BOOSTROOTRELPATH
+        mkdir -p tmp && cd tmp
+        if [ ! -f saxonhe.zip ]; then curl -s -S --retry 10 -L -o saxonhe.zip https://sourceforge.net/projects/saxon/files/Saxon-HE/9.9/SaxonHE9-9-1-4J.zip/download; fi
+        unzip -d saxonhe -o saxonhe.zip
+        cd saxonhe
 
-    sudo rm /Library/Java/Extensions/Saxon-HE.jar || true
-    sudo cp saxon9he.jar /Library/Java/Extensions/Saxon-HE.jar
-    sudo chmod 755 /Library/Java/Extensions/Saxon-HE.jar
+        sudo rm /Library/Java/Extensions/Saxon-HE.jar || true
+        sudo cp saxon9he.jar /Library/Java/Extensions/Saxon-HE.jar
+        sudo chmod 755 /Library/Java/Extensions/Saxon-HE.jar
+    fi
 
 fi
 
 # In the above 'packages' section a python virtenv was created. Activate it, if that has not been done already.
 if [ -f ${pythonvirtenvpath}/bin/activate ]; then
     source ${pythonvirtenvpath}/bin/activate
+fi
+
+# In the above 'packages' section npm was installed. Activate it, if that has not been done already.
+if [ -d $HOME/.nvm_${REPONAME}_antora ]; then
+        export NODE_VERSION=18.18.1
+        # The container has a pre-installed nodejs. Overwrite those again.
+        export NVM_BIN="$HOME/.nvm_${REPONAME}_antora/versions/node/v${NODE_VERSION}/bin"
+        export NVM_DIR=$HOME/.nvm_${REPONAME}_antora
+        export NVM_INC=$HOME/.nvm_${REPONAME}_antora/versions/node/v${NODE_VERSION}/include/node
+        . "$NVM_DIR/nvm.sh" && nvm use v${NODE_VERSION}
+        export PATH="$HOME/.nvm_${REPONAME}_antora/versions/node/v${NODE_VERSION}/bin/:${PATH}"
+        node --version
+        npm --version
 fi
 
 # check this on apple silicon:
@@ -355,7 +392,7 @@ if [ -d ${BOOST_ROOT}/build/docbook-xml ]; then
     export DOCBOOK_DTD_DIR=${BOOST_ROOT}/build/docbook-xml
 fi
 
-if [ "$skipboostoption" != "yes" ] ; then
+if [ "$skipboostoption" != "yes" ] && [ "$typeoption" != "antora" ] ; then
 
     git submodule update --init libs/context
     git submodule update --init tools/boostbook
@@ -417,7 +454,7 @@ if [ ! -d $librarypath/doc ]; then
     exit 0
 fi
 
-if [ -f $librarypath/doc/Jamfile ] || [ -f $librarypath/doc/jamfile ] || [ -f $librarypath/doc/Jamfile.v2 ] || [ -f $librarypath/doc/jamfile.v2 ] || [ -f $librarypath/doc/Jamfile.v3 ] || [ -f $librarypath/doc/jamfile.v3 ] || [ -f $librarypath/doc/Jamfile.jam ] || [ -f $librarypath/doc/jamfile.jam ] || [ -f $librarypath/doc/build.jam ] ; then
+if [ -f $librarypath/doc/build_antora.sh ] || [ -f $librarypath/doc/Jamfile ] || [ -f $librarypath/doc/jamfile ] || [ -f $librarypath/doc/Jamfile.v2 ] || [ -f $librarypath/doc/jamfile.v2 ] || [ -f $librarypath/doc/Jamfile.v3 ] || [ -f $librarypath/doc/jamfile.v3 ] || [ -f $librarypath/doc/Jamfile.jam ] || [ -f $librarypath/doc/jamfile.jam ] || [ -f $librarypath/doc/build.jam ] ; then
      : # ok
 else
     echo "doc/Jamfile or similar is missing for this library. No need to compile. Exiting."
@@ -449,14 +486,25 @@ if [ "$typeoption" = "main" ]; then
     echo "using quickbook : build/dist/bin/quickbook ; using auto-index : build/dist/bin/auto_index ; using docutils ; using doxygen ; using boostbook ; using asciidoctor ; using saxonhe ;" > tools/build/src/user-config.jam
     ./b2 -j3 $librarypath/doc${boostrelease}
 
-elif  [ "$typeoption" = "cppalv1" ]; then
+elif [ "$typeoption" = "antora" ]; then
+    cd ${librarypath}/doc
+    chmod 755 build_antora.sh
+    ./build_antora.sh
+
+elif [ "$typeoption" = "cppalv1" ]; then
     echo "using doxygen ; using boostbook ; using saxonhe ;" > tools/build/src/user-config.jam
     ./b2 $librarypath/doc${boostrelease} cxxstd=11
 fi
 
+if [ "$typeoption" = "antora" ]; then
+    result_sub_path="doc/build/site/"
+else
+    result_sub_path="doc/html/"
+fi
+
 if [ "${BOOSTROOTLIBRARY}" = "yes" ]; then
     echo ""
-    echo "Build completed. View the results in $librarypath/doc/html"
+    echo "Build completed. View the results in $librarypath/${result_sub_path}"
     echo ""
 else
     if  [ "$BOOSTROOTRELPATH" = "." ]; then
@@ -465,6 +513,6 @@ else
         pathfiller="/${BOOSTROOTRELPATH}/"
     fi
     echo ""
-    echo "Build completed. View the results in ${BOOST_SRC_FOLDER}${pathfiller}boost-root/$librarypath/doc/html"
+    echo "Build completed. View the results in ${BOOST_SRC_FOLDER}${pathfiller}boost-root/$librarypath/${result_sub_path}"
     echo ""
 fi
